minor changes to psoc driver for sending the message, not tested

+++ rover_ws/src/psoc_driver/include/psoc.h
@@ -25,7 +25,7 @@ public:
     bool received;
     std::stringstream out;
     /* void send(std::string command); */
-    void send(uint8_t, uint8_t, uint8_t, uint8_t, uint8_t);
+    void send(uint16_t, uint16_t, uint16_t, uint16_t, uint8_t);
 
 private:
     ros::NodeHandle nh_;
