+++ rover_ws/src/psoc_driver/src/psoc.cpp
@@ -57,7 +57,7 @@ void Psoc::receive(const uint8_t *bytes, ssize_t nbytes)
 //     link->send_bytes((uint8_t*)data, length);
 // }
 
-void Psoc::send(uint8_t lw, uint8_t rw, uint8_t pan, uint8_t tilt,uint8_t camnum)
+void Psoc::send(uint16_t lw, uint16_t rw, uint16_t pan, uint16_t tilt, uint8_t camnum)
 {
   uint8_t array[11];
   array[0]=0xEA;
